[fix] Calcs am/pm correctly when hours is > 24
In `struct tm`, it is valid to encode an "improper" time in order to
describe a delta relative to a proper time, without manually performing
the calendar manipulations necessary to encode it properly. As an example,
it is valid to start from 11pm (23h) and add two hours, relying on the
`struct tm` handling code to normalize this. However, the time formatting
code in this library did not handle the case where `hours > 24` correctly,
interpreting it as merely an am/pm switch.